requirements.txt
#
# Requirements for caspython-centera
#
nose
python-dateutil
